v0.2.3.sql

+++ upgrade/sql/v0.2.3.sql
@@ -7,7 +7,7 @@ ALTER TABLE `pay_trade_chan`
 CHANGE COLUMN `chan_data` `pay_code` VARCHAR(40) CHARACTER SET 'utf8' COLLATE 'utf8_unicode_ci' NOT NULL COMMENT '支付代码' AFTER `pay_method`,
 CHANGE COLUMN `trade_no` `trade_no` VARCHAR(40) CHARACTER SET 'utf8' COLLATE 'utf8_unicode_ci' NOT NULL COMMENT '支付单号' ,
 CHANGE COLUMN `pay_chan` `pay_method` INT(6) NOT NULL COMMENT '支付方式' ,
-CHANGE COLUMN `internal_chan` `internal` TINYINT(1) NOT NULL COMMENT '是否为内置支付方式' ,
+CHANGE COLUMN `internal_chan` `internal` INT(1) NOT NULL COMMENT '是否为内置支付方式' ,
 ADD COLUMN `out_trade_no` VARCHAR(45) NOT NULL COMMENT '外部交易单号' AFTER `pay_amount`, RENAME TO  `pay_trade_data` ;
 
 ALTER TABLE `pay_trade_data`
